Completed LocationDB Controllers
I still need to edit the FXML to adjust for added/changed names/methods, but the controllers should be good. Hopefully should be good for presentation. I also don't have FX on chromebook, so that needs to be tested with all the right packages.

I modeled them after Calvin's InstructorDB controllers.